📦️ Modernize Dockerfiles to support IPv6
We want to be able to run on IPv6-only Kubernetes clusters.
Currently, multiple components bind directly to 0.0.0.0,
which excludes IPv6 connections. Let's change that, and
make it possible to change bind address from the Dockerfile
as well; and bump up the very old Node version while we're
here. Also, put code in /app instead of /.